Refactor javaScript files - Add location.js, render.js, search.js, and weather.js in mymodules
